fix Prettier crash from multiple cursors in .vue files
Prettier crashed when formatting .vue files due to multiple 'cursor' markers being inserted in the generated doc. This happens when both nodeBeforeCursor and nodeAfterCursor are set, but no actual cursor is present in the input.

This patch ensures only one cursor is inserted by tracking it with a _cursorApplied flag.

Signed-off-by: Toni Förster <toni.foerster@icloud.com>